Associative choice with identity.
This corrects the `L.firstOf` mistake where I didn't see the possibility
of a suitable identity element for choice.  `L.nothing` acts like the
identity element of `L.choice`.  `L.orElse` can be used to give a backup
or default if it is really needed, but none of the actual use cases I've
seen so far require a backup.