Change semantics: empty match terminates traversal
By making an empty match terminate the traversal, the possibility of going to an
infinite loop is eliminated.  This behaviour could also be useful: You could
craft a regular expression that produces and empty match intentionally to
terminate the traversal.  The warning given in non-production builds reflects
that often this is caused an incorrect regular expression than unintentionally
can produce empty matches.